+++ src/docker/pull.js
@@ -36,20 +36,46 @@ function parse_status(msg) {
   return result;
 }
 
-export function pull(docker, repository, tag, stdout) {
+export function pull(docker, repository, tag, stdout, registry_result) {
   var image   = `${repository}:${tag}`;
   var promise = docker.createImage({
     fromImage: repository,
     tag: tag,
   });
 
+  var bar = null;
+  var registry_layers_ids_count = 0;
+  // var non_existent_locally_ids_count = 0;
+  // var total_layer_size_left = 0;
+  var will_show_simple_progress_bar = false;
+
+  if (!_.isNull(registry_result)) {
+    will_show_simple_progress_bar = true;
+    registry_layers_ids_count      = registry_result.registry_layers_ids_count;
+    // non_existent_locally_ids_count = registry_result.non_existent_locally_ids_count;
+    // total_layer_size_left          = registry_result.total_layer_size_left;
+  }
+
+  if (will_show_simple_progress_bar) {
+    var ProgressBar = require('progress');
+    var progressMessage = ' [:bar] :percent';
+    bar = new ProgressBar(progressMessage, {
+      complete: '=',
+      incomplete: ' ',
+      width: 23,
+      total: registry_layers_ids_count + 1
+    });
+  }
+
   return promise.then((stream) => {
     return defer((resolve, reject, notify) => {
       stream.on('data', (data) => {
         // TODO: add support chucks
         try {
+
           var msg  = JSON.parse(data.toString());
           msg.type = "pull_msg";
+
           if (msg.error) {
             if (msg.error.match(/404/) || msg.error.match(/not found$/)) {
               return reject(new ProvisionNotFound(image));
@@ -57,7 +83,14 @@ export function pull(docker, repository, tag, stdout) {
             reject(new ProvisionPullError(image, msg.error));
           } else {
             msg.statusParsed = parse_status(msg.status);
-            if (msg.statusParsed) {
+
+            if ( will_show_simple_progress_bar &&
+                msg.statusParsed &&
+                msg.statusParsed.type === 'download_complete') {
+              // show a simple progress-bar
+              bar.tick(1);
+            } else if (!will_show_simple_progress_bar && msg.statusParsed) {
+              // show messages from docker remote pull
               notify(msg);
             }
             if (stdout) {